+++ README.md
@@ -62,7 +62,7 @@
     - [x] 행동 : 상하좌우 1칸으로 이동할 수 있는 지점 여부 반환
     - [x] 행동 : 이동 지점에 대한 이동 경로를 반환함
     - [x] 행동 : 이동 경로에 장애물이 없어 실제로 이동 가능한지 여부 반환
-  - [ ] 마
+  - [X] 마
     - [x] 색상(팀)
     - [x] 행동 : 상하좌우 1칸 + 대각선 1칸으로 이동할 수 있는 지점 여부 반환
     - [x] 행동 : 이동 지점에 대한 이동 경로를 반환함
