+++ src/main/java/model/janggiboard/JanggiBoard.java
@@ -1,6 +1,6 @@
-package model.janggi_board;
+package model.janggiboard;
 
-import static model.janggi_board.JanggiBoardSetUp.DEFAULT_SETUP;
+import static model.janggiboard.JanggiBoardSetUp.DEFAULT_SETUP;
 
 import java.util.ArrayList;
 import java.util.HashMap;
